app.py
import streamlit as st
from PIL import Image
import numpy as np
import torch
import timm
import cv2

st.set_page_config(page_title="ML Image Overlay", layout="wide")

st.title("ML Image Overlay Demo")

# Upload an image
uploaded_file = st.file_uploader("Upload an image", type=["jpg", "png", "jpeg"])
if uploaded_file:
    image = Image.open(uploaded_file).convert("RGB")
    st.image(image, caption="Original Image", use_column_width=True)

    # Convert to numpy array for overlay processing
    img_array = np.array(image)

    # --- Dummy ML inference ---
    # Replace this with your actual model inference
    st.write("Running dummy ML inference...")
    overlay = img_array.copy()
    overlay[:, :, 0] = 255 - overlay[:, :, 0]  # Just invert the red channel as a placeholder

    # Convert back to Image for display
    overlay_image = Image.fromarray(overlay)
    st.image(overlay_image, caption="Overlay Result", use_column_width=True)

    st.success("Done! Replace the dummy ML code with your model.")
