requirements.txt
streamlit==1.39.0
torch==2.2.0
timm==1.7.2
numpy
Pillow
opencv-python
